requirements.txt
# Maily — requirements
cerebras-cloud-sdk>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
networkx>=3.0
matplotlib>=3.7.0
streamlit>=1.35.0
langgraph